필수 기능 수정, 리팩토링

+++ src/main/java/com/github/gajicoding/schedule_api_project/api/v1/service/impl/ScheduleCommentServiceImpl.java
@@ -5,9 +5,9 @@ import com.github.gajicoding.schedule_api_project.api.v1.data.dto.schedule_comme
 import com.github.gajicoding.schedule_api_project.api.v1.data.entity.Schedule;
 import com.github.gajicoding.schedule_api_project.api.v1.data.entity.ScheduleComment;
 import com.github.gajicoding.schedule_api_project.api.v1.data.entity.User;
-import com.github.gajicoding.schedule_api_project.api.v1.exception.ScheduleCommentExceptions;
-import com.github.gajicoding.schedule_api_project.api.v1.exception.ScheduleExceptions;
-import com.github.gajicoding.schedule_api_project.api.v1.exception.UserExceptions;
+import com.github.gajicoding.schedule_api_project.api.v1.exception.factory.ScheduleCommentExceptionFactory;
+import com.github.gajicoding.schedule_api_project.api.v1.exception.factory.ScheduleExceptionFactory;
+import com.github.gajicoding.schedule_api_project.api.v1.exception.factory.UserExceptionFactory;
 import com.github.gajicoding.schedule_api_project.api.v1.repository.ScheduleCommentRepository;
 import com.github.gajicoding.schedule_api_project.api.v1.repository.ScheduleRepository;
 import com.github.gajicoding.schedule_api_project.api.v1.repository.UserRepository;
@@ -27,8 +27,8 @@ public class ScheduleCommentServiceImpl implements ScheduleCommentService {
 
     @Override
     public ScheduleCommentResponseDTO save(Long scheduleId, Long userId, ScheduleCommentRequestDTO requestDTO) {
-        Schedule schedule = scheduleRepository.findById(scheduleId).orElseThrow(()-> ScheduleExceptions.notFoundById(scheduleId));
-        User user = userRepository.findById(userId).orElseThrow(() -> UserExceptions.notFoundById(userId));
+        Schedule schedule = scheduleRepository.findById(scheduleId).orElseThrow(()-> ScheduleExceptionFactory.notFoundById(scheduleId));
+        User user = userRepository.findById(userId).orElseThrow(() -> UserExceptionFactory.notFoundById(userId));
 
         ScheduleComment scheduleComment = new ScheduleComment(requestDTO.getContents(), user, schedule);
 
@@ -37,12 +37,12 @@ public class ScheduleCommentServiceImpl implements ScheduleCommentService {
 
     @Override
     public ScheduleCommentResponseDTO findById(Long id) {
-        return new ScheduleCommentResponseDTO(scheduleCommentRepository.findById(id).orElseThrow(()-> ScheduleCommentExceptions.notFoundById(id)));
+        return new ScheduleCommentResponseDTO(scheduleCommentRepository.findById(id).orElseThrow(()-> ScheduleCommentExceptionFactory.notFoundById(id)));
     }
 
     @Override
     public List<ScheduleCommentResponseDTO> findAll(Long scheduleId) {
-        return scheduleCommentRepository.findAll()
+        return scheduleCommentRepository.findAllBySchedule_Id(scheduleId)
                 .stream()
                 .map(ScheduleCommentResponseDTO::new)
                 .toList();
@@ -50,15 +50,25 @@ public class ScheduleCommentServiceImpl implements ScheduleCommentService {
 
     @Transactional
     @Override
-    public ScheduleCommentResponseDTO updateContents(Long id, ScheduleCommentRequestDTO requestDTO) {
-        ScheduleComment scheduleComment = scheduleCommentRepository.findById(id).orElseThrow(()-> ScheduleCommentExceptions.notFoundById(id));
+    public ScheduleCommentResponseDTO updateContents(Long id, Long userId, ScheduleCommentRequestDTO requestDTO) {
+        ScheduleComment scheduleComment = scheduleCommentRepository.findById(id).orElseThrow(()-> ScheduleCommentExceptionFactory.notFoundById(id));
+
+        if(!userId.equals(scheduleComment.getUser().getId())){
+            throw ScheduleCommentExceptionFactory.noPermissionToUpdate();
+        }
+
         scheduleComment.setContents(requestDTO.getContents());
         return new ScheduleCommentResponseDTO(scheduleComment);
     }
 
     @Override
-    public void delete(Long id) {
-        ScheduleComment scheduleComment = scheduleCommentRepository.findById(id).orElseThrow(()-> ScheduleCommentExceptions.notFoundById(id));
+    public void delete(Long id, Long userId) {
+        ScheduleComment scheduleComment = scheduleCommentRepository.findById(id).orElseThrow(()-> ScheduleCommentExceptionFactory.notFoundById(id));
+
+        if(!userId.equals(scheduleComment.getUser().getId())){
+            throw ScheduleCommentExceptionFactory.noPermissionToDelete();
+        }
+
         scheduleCommentRepository.delete(scheduleComment);
     }
 }
